requirements.txt
youtube-transcript-api
openai
python-dotenv
streamlit
reportlab
yt_dlp